be/requirements.txt
Django==5.1.1 
django-cors-headers==4.4.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
djangorestframework==3.15.2
gunicorn==23.0.0
Pillow==10.4.0